backend/pyproject.toml
[project]
name = "nova-devops-copilot-backend"
version = "0.1.0"
description = "Amazon Nova DevOps Copilot — 4-agent pipeline backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "boto3>=1.34.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
